use Commander and make the whole thing with CLI args
