--- .env
@@ -1,2 +1,0 @@
-WP_SITE_A=example-a.com
-WP_SITE_B=example-b.com
