+++ package.json
@@ -9,9 +9,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "command-line-args": "^5.0.2",
+    "commander": "2.17.1",
     "deep-diff": "^1.0.1",
-    "dotenv": "^6.0.0",
     "request": "^2.87.0",
     "winston": "^3.0.0"
   }
